Refactor test utilities (#756)
