+++ test/common_utils/parameterized_utils.py
@@ -1,0 +1,10 @@
+import json
+
+from parameterized import param
+
+from .data_utils import get_asset_path
+
+
+def load_params(*paths):
+    with open(get_asset_path(*paths), 'r') as file:
+        return [param(json.loads(line)) for line in file]
