+++ test/sox_io_backend/test_load.py
@@ -8,14 +8,14 @@ from ..common_utils import (
     PytorchTestCase,
     skipIfNoExec,
     skipIfNoExtension,
-)
-from .common import (
-    get_test_name,
     get_wav_data,
     load_wav,
     save_wav,
+    sox_utils,
 )
-from . import sox_utils
+from .common import (
+    name_func,
+)
 
 
 class LoadTestBase(TempDirMixin, PytorchTestCase):
@@ -129,7 +129,7 @@ class TestLoad(LoadTestBase):
         [8000, 16000],
         [1, 2],
         [False, True],
-    )), name_func=get_test_name)
+    )), name_func=name_func)
     def test_wav(self, dtype, sample_rate, num_channels, normalize):
         """`sox_io_backend.load` can load wav format correctly."""
         self.assert_wav(dtype, sample_rate, num_channels, normalize, duration=1)
@@ -139,7 +139,7 @@ class TestLoad(LoadTestBase):
         [16000],
         [2],
         [False],
-    )), name_func=get_test_name)
+    )), name_func=name_func)
     def test_wav_large(self, dtype, sample_rate, num_channels, normalize):
         """`sox_io_backend.load` can load large wav file correctly."""
         two_hours = 2 * 60 * 60
@@ -148,7 +148,7 @@ class TestLoad(LoadTestBase):
     @parameterized.expand(list(itertools.product(
         ['float32', 'int32', 'int16', 'uint8'],
         [4, 8, 16, 32],
-    )), name_func=get_test_name)
+    )), name_func=name_func)
     def test_multiple_channels(self, dtype, num_channels):
         """`sox_io_backend.load` can load wav file with more than 2 channels."""
         sample_rate = 8000
@@ -159,7 +159,7 @@ class TestLoad(LoadTestBase):
         [8000, 16000, 44100],
         [1, 2],
         [96, 128, 160, 192, 224, 256, 320],
-    )), name_func=get_test_name)
+    )), name_func=name_func)
     def test_mp3(self, sample_rate, num_channels, bit_rate):
         """`sox_io_backend.load` can load mp3 format correctly."""
         self.assert_mp3(sample_rate, num_channels, bit_rate, duration=1)
@@ -168,7 +168,7 @@ class TestLoad(LoadTestBase):
         [16000],
         [2],
         [128],
-    )), name_func=get_test_name)
+    )), name_func=name_func)
     def test_mp3_large(self, sample_rate, num_channels, bit_rate):
         """`sox_io_backend.load` can load large mp3 file correctly."""
         two_hours = 2 * 60 * 60
@@ -178,7 +178,7 @@ class TestLoad(LoadTestBase):
         [8000, 16000],
         [1, 2],
         list(range(9)),
-    )), name_func=get_test_name)
+    )), name_func=name_func)
     def test_flac(self, sample_rate, num_channels, compression_level):
         """`sox_io_backend.load` can load flac format correctly."""
         self.assert_flac(sample_rate, num_channels, compression_level, duration=1)
@@ -187,7 +187,7 @@ class TestLoad(LoadTestBase):
         [16000],
         [2],
         [0],
-    )), name_func=get_test_name)
+    )), name_func=name_func)
     def test_flac_large(self, sample_rate, num_channels, compression_level):
         """`sox_io_backend.load` can load large flac file correctly."""
         two_hours = 2 * 60 * 60
@@ -197,7 +197,7 @@ class TestLoad(LoadTestBase):
         [8000, 16000],
         [1, 2],
         [-1, 0, 1, 2, 3, 3.6, 5, 10],
-    )), name_func=get_test_name)
+    )), name_func=name_func)
     def test_vorbis(self, sample_rate, num_channels, quality_level):
         """`sox_io_backend.load` can load vorbis format correctly."""
         self.assert_vorbis(sample_rate, num_channels, quality_level, duration=1)
@@ -206,7 +206,7 @@ class TestLoad(LoadTestBase):
         [16000],
         [2],
         [10],
-    )), name_func=get_test_name)
+    )), name_func=name_func)
     def test_vorbis_large(self, sample_rate, num_channels, quality_level):
         """`sox_io_backend.load` can load large vorbis file correctly."""
         two_hours = 2 * 60 * 60
@@ -230,14 +230,14 @@ class TestLoadParams(TempDirMixin, PytorchTestCase):
     @parameterized.expand(list(itertools.product(
         [0, 1, 10, 100, 1000],
         [-1, 1, 10, 100, 1000],
-    )), name_func=get_test_name)
+    )), name_func=name_func)
     def test_frame(self, frame_offset, num_frames):
         """num_frames and frame_offset correctly specify the region of data"""
         found, _ = sox_io_backend.load(self.path, frame_offset, num_frames)
         frame_end = None if num_frames == -1 else frame_offset + num_frames
         self.assertEqual(found, self.original[:, frame_offset:frame_end])
 
-    @parameterized.expand([(True, ), (False, )], name_func=get_test_name)
+    @parameterized.expand([(True, ), (False, )], name_func=name_func)
     def test_channels_first(self, channels_first):
         """channels_first swaps axes"""
         found, _ = sox_io_backend.load(self.path, channels_first=channels_first)
